By the power of Docker, I compel you!
This PR changes the doc generator script to use Docker images for all
the weird utilities. That includes Thrift itself, so now we have
explicit control over which version of Thrift is used to generate the
docs. I opted for the latest one, which seems to have some minor
differences in the generated HTML. FWIW, I prefer the new look.

This PR also regenerates the docs.

It also downgrades tidy by 5 patch versions. Whatevs.